feat(core): update assertDomNode to catch the case where a node comes from another frame
Previously if this function is used on a DOM node that is created by a frame other than the main one, it'll fail this assert.

A debug function `getDirectives`, eventually calls `assertDomNode` in its call tree. This caused an issue with Angular DevTools users who use third party scripts that:
1. creates an iframe
2. uses the iframes document to create a DOM element
3. injects that DOM element into the main document

This commit attempts to resolve this issue by augmenting assertDomNode to account for the case where a DomNode comes from another frame. Unfortunately there is no definitive way of doing this without having access to the frame in question, so we instead use a mostly accurate heuristic.