feat(service-worker): allow specifying maxAge for entire application
This commit adds an `applicationMaxAge` to the service worker configuration. When set, it will only return cached responses to asset, data and navigation requests within the maxAge. Afterwards, it will always fetch from the network for all requests, if possible. The default is `undefined`, for which the behaviour is the same as it currently is.