Use TaskType keys for counting tasks instead of hardcoding the key values
Change the type of `_taskCounts` to an IndexSignature that can have keys
named as the values of the `TaskType` type.

The Closure Compiler used at Google has a property renaming optimization
that can change the property names when minifying code. Having the
correct type helps the TSJS team that develops a tool to identfy
property renaming issues directly in TypeScript.

Signed-off-by: Costin Sin <sin.costinrobert@gmail.com>