feat(elements): support signal-based components
Support signal-based components in createCustomElement.

Previously, the signal was overwritten by the input value, losing
reactivity. This change calls setInput on the componentRef.

Also remove onChanges logic in ComponentFactoryStrategy,
because setInput handles this.

DEPRECATED: the injector argument of ComponentNgElementStrategyFactory

The injector argument is no longer needed in the constructor.

Fixes #53981