refactor(compiler): add `i18nPreserveWhitespaceForLegacyExtraction`
This configures whether or not to preserve whitespace content when extracting messages from Angular templates in the legacy (View Engine) extraction pipeline.

The change to `i18n_parser.ts` is really a bug fix. The localization process requires that the nubmer of tokens in a message stays constant between i18n passes so it can reuse source spans. Therefore even empty text tokens need to be retained in order to maintain that invariant. However this bug fix does affect message IDs, and changing that could invalidate existing translations, therefore it is also gated by the same flag to maintain backwards compatibility.